requirements.txt
streamlit==1.39.0
pandas==2.0.3
google-auth==2.22.0
google-auth-oauthlib==0.4.4
google-auth-httplib2==0.1.0
google-api-python-client==2.23.0